fixed various warnings

+++ src/MessageReader.cpp
@@ -11,7 +11,7 @@ uint8_t read_message(struct message_t *message) {
 	switch (message->state) {
 		case WAITING_FOR_MESSAGE: {
 			if (Serial.available() > 0) {
-				message->data.unorganizedMessage[message->data.length] = Serial.read();
+				message->data.unorganizedMessage[message->data.length] =(uint8_t)Serial.read();
 				message->data.length++;
 				if (message->data.unorganizedMessage[message->data.length - 1] == 0xD){
