+++ src/main.cpp
@@ -25,10 +25,10 @@ int main(){
 void setup(void) {
   Serial.begin(9600);
 
-	for(int i=47; i>=41; i-=2){ //rows
+	for(uint8_t i=47; i>=41; i-=2){ //rows
 		pinMode(i, OUTPUT);
 	}
-	for(int i=39; i>=33; i-=2){ //columns
+	for(uint8_t i=39; i>=33; i-=2){ //columns
 		pinMode(i, INPUT_PULLUP);
 	}
 
@@ -38,7 +38,7 @@ void setup(void) {
 	TCCR3B = 0;// same for TCCR1B
 	TCNT3  = 0;//initialize counter value to 0
 	// set compare match register for 1hz increments
-	OCR3A = F_CPU / 200;// = (16*10^6) / (1*1024) - 1 (must be <65536)
+	OCR3A = (uint16_t)(F_CPU / 200);// = (16*10^6) / (1*1024) - 1 (must be <65536)
 	// turn on CTC mode
 	TCCR3B |= (1 << WGM32);
 	// Set CS10 and CS12 bits for 1024 prescaler
